Fix WM_CLASS woes
Upstream Firefox Nightly build has the app-id `firefox-nightly` so KDE
expects the desktop file to have the same name. But the desktop file name
comes from the $FLATPAK_ID which is different from the app-id.

So start firefox with the $FLATPAK_ID as app-id and change the
StartupWMClass to match that

Ref. https://gitlab.com/projects261/firefox-dev-flatpak/-/issues/1